Single partition performance test.

+++ samza/common/src/main/java/edu/zhangfan/cs237/common/DriverLocationEvent.java
@@ -3,11 +3,11 @@ package edu.zhangfan.cs237.common;
 public class DriverLocationEvent {
   private Integer blockId;
   private String driverId;
-  private Integer latitude;
-  private Integer longitude;
+  private double latitude;
+  private double longitude;
   private Type type;
 
-  public DriverLocationEvent(Integer blockId, String driverId, Integer latitude, Integer longitude, Type type) {
+  public DriverLocationEvent(Integer blockId, String driverId, double latitude, double longitude, Type type) {
     this.blockId = blockId;
     this.driverId = driverId;
     this.latitude = latitude;
@@ -31,19 +31,19 @@ public class DriverLocationEvent {
     this.driverId = driverId;
   }
 
-  public Integer getLatitude() {
+  public double getLatitude() {
     return latitude;
   }
 
-  public void setLatitude(Integer latitude) {
+  public void setLatitude(double latitude) {
     this.latitude = latitude;
   }
 
-  public Integer getLongitude() {
+  public double getLongitude() {
     return longitude;
   }
 
-  public void setLongitude(Integer longitude) {
+  public void setLongitude(double longitude) {
     this.longitude = longitude;
   }
